requirements.txt
# primary package - AWS bindings 
aws-cdk-lib
argparse
python-dotenv
rich
boto3

